refactor: break Syncing component into chunks
