+++ renderer/components/Syncing/index.js
@@ -1,1 +1,8 @@
-export Syncing from './Syncing'
+import Syncing from './Syncing'
+
+export default Syncing
+export Address from './Address'
+export NewWalletHeader from './NewWalletHeader'
+export OldWalletHeader from './OldWalletHeader'
+export Progress from './Progress'
+export Tutorials from './Tutorials'
