+++ renderer/containers/Syncing.js
@@ -6,7 +6,7 @@ import { infoSelectors } from 'reducers/info'
 import { neutrinoSelectors } from 'reducers/neutrino'
 import { setIsWalletOpen } from 'reducers/wallet'
 import { showNotification } from 'reducers/notification'
-import { Syncing } from 'components/Syncing'
+import Syncing from 'components/Syncing'
 import { Modal, ModalOverlayStyles } from 'components/UI'
 import { useOnKeydown } from 'hooks'
